requirements.txt
<<<<<<< HEAD
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
datasets==4.8.4
deep-translator==1.11.4
dill==0.4.1
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.0
idna==3.11
Jinja2==3.1.6
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pdf2image==1.17.0
pillow==12.2.0
propcache==0.4.1
psutil==7.2.2
pyarrow==23.0.1
Pygments==2.20.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.1
rich==14.3.3
safetensors==0.7.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
sympy==1.14.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
xxhash==3.6.0
yarl==1.23.0
=======
streamlit
plotly
python-dotenv
edge-tts
google-genai
httpx
PyPDF2
beautifulsoup4
>>>>>>> dd65b8b0ff520f33678bff0ce96fb968e9ed8166
